CHROMIUM: Use File for device request fd
There is nothing in the vhost-user documentation or spec that says this
must be a socket for fs device requests. It's not even guaranteed that
it is an fd for a UnixStream.

Change the type to a File and leave it up to the backend to convert it
into a more specialized type if necessary.

Upstream PR: https://github.com/rust-vmm/vhost/pull/49

BUG=b:179755651
TEST=cargo test

Cq-Depend: chromium:2987594
Change-Id: I52ce0c62e0627a87462be8e0a44fa534763f10ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/rust-vmm/vhost/+/2988140
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>